Blanket format string fixes from previous commit.

+++ AppController.m
@@ -2997,7 +2997,7 @@ void outletObjectAwoke(id sender) {
             NSUInteger theCount = [[[textView textStorage] words] count];
 
             if (theCount > 0) {
-                [wordCounter setStringValue:[[NSString stringWithFormat:@"%d", theCount] stringByAppendingString:@" words"]];
+                [wordCounter setStringValue:[[NSString stringWithFormat:@"%lu", (unsigned long)theCount] stringByAppendingString:@" words"]];
             }else {
                 [wordCounter setStringValue:@""];
             }
